+++ bin/gensamples.dart
@@ -37,11 +37,15 @@
 //     THE SOFTWARE.
 
 import 'dart:io';
+
 import 'package:path/path.dart';
 import 'package:json_schema/json_schema.dart';
+import 'package:json_schema/vm.dart';
 import 'package:json_schema/schema_dot.dart';
 
 main() {
+  configureJsonSchemaForVm();
+
   final sourcePath = join(dirname(dirname(absolute(Platform.script.toFilePath()))), 'dot_samples', 'schemas');
   final outPath = join(dirname(sourcePath), 'schemaout');
   new Directory(sourcePath).listSync().forEach((jsonFile) {
@@ -51,6 +55,7 @@ main() {
     final pngOut = join(outPath, '$base.png');
 
     JsonSchema.createSchemaFromUrl(fname).then((schema) {
+      schema.refMap.forEach((key, ref) => print('$key : $ref'));
       new File(dotFilename).writeAsStringSync(createDot(schema));
     }).then((_) {
       Process.run('dot', ['-Tpng', '-o$pngOut', dotFilename]).then((ProcessResult processResult) {