+++ example/from_url/validate_instance_from_url.dart
@@ -38,6 +38,7 @@
 //     THE SOFTWARE.
 
 import 'package:dart2_constant/convert.dart' as convert;
+
 import 'package:json_schema/json_schema.dart';
 import 'package:logging/logging.dart';
 
@@ -50,12 +51,6 @@ main() {
   //////////////////////////////////////////////////////////////////////
   String url = 'http://json-schema.org/draft-04/schema';
   JsonSchema.createSchemaFromUrl(url).then((JsonSchema schema) {
-    // TODO: Figure out the redirect issues here
-    //   if(false) {
-    //     print('''Does schema validate itself?
-    //       ${schema.validate(schema.schemaMap)}''');
-    //   }
-
     final validSchema = {'type': 'integer'};
     print('''Does schema validate valid schema $validSchema?
   ${schema.validate(validSchema)}''');