+++ example/readme/asynchronous_creation/remote_http_refs.dart
@@ -1,3 +1,4 @@
+#!/usr/bin/env dart
 // Copyright 2013-2018 Workiva Inc.
 //
 // Licensed under the Boost Software License (the "License");
@@ -36,8 +37,42 @@
 //     OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 //     THE SOFTWARE.
 
-export 'package:json_schema/src/json_schema/json_schema.dart' show JsonSchema;
-export 'package:json_schema/src/json_schema/constants.dart' show SchemaVersion;
-export 'package:json_schema/src/json_schema/schema_type.dart' show SchemaType;
-export 'package:json_schema/src/json_schema/validator.dart' show Validator;
-export 'package:json_schema/src/json_schema/typedefs.dart' show RefProvider, RefProviderAsync;
+import 'dart:io';
+
+import 'package:json_schema/json_schema.dart';
+
+// For VM:
+import 'package:json_schema/vm.dart';
+
+// For Browser:
+// import 'package:json_schema/browser.dart';
+
+main() async {
+  // For VM:
+  configureJsonSchemaForVm();
+
+  // For Browser:
+  // configureJsonSchemaForBrowser();
+
+  // Schema Defined as a JSON String
+  final schema = await JsonSchema.createSchemaAsync(r'''
+  {
+    "type": "array",
+    "items": {
+      "$ref": "https://raw.githubusercontent.com/json-schema-org/JSON-Schema-Test-Suite/master/remotes/integer.json"
+    }
+  }
+  ''');
+
+  // Create some examples to validate against the schema.
+  final numbersArray = [1, 2, 3];
+  final decimalsArray = [3.14, 1.2, 5.8];
+  final strArray = ['hello', 'world'];
+
+  print('$numbersArray => ${schema.validate(numbersArray)}'); // true
+  print('$decimalsArray => ${schema.validate(decimalsArray)}'); // false
+  print('$strArray => ${schema.validate(strArray)}'); // false
+
+  // Exit the process cleanly (VM Only).
+  exit(0);
+}