+++ lib/src/json_schema/type_validators.dart
@@ -40,7 +40,7 @@ class TypeValidators {
     throw FormatExceptions.error('$key must be a non-empty string: $value');
   }
 
-  static List<SchemaType> schemaTypeList(String key, dynamic value) {
+  static List<SchemaType> typeList(String key, dynamic value) {
     var typeList;
     if (value is String) {
       typeList = [SchemaType.fromString(value)];
@@ -84,10 +84,11 @@ class TypeValidators {
     throw FormatExceptions.object(key, value);
   }
 
-  static String jsonSchemaVersion4Or6(String key, dynamic value) {
+  static SchemaVersion jsonSchemaVersion4Or6(String key, dynamic value) {
     string(key, value);
-    if (JsonSchemaVersions.allVersions.contains(value)) {
-      return value;
+    final schemaVersion = SchemaVersion.fromString(value);
+    if (schemaVersion != null) {
+      return schemaVersion;
     }
     throw FormatExceptions.error('Only draft 4 and draft 6 schemas supported');
   }